app/requirements.txt
Flask==2.0.2
mysql_connector_repackaged==0.3.1
requests
mysql-connector-python==8.0.17
flask-mysqldb

pyyaml

